fold InferWidthsWithAnnos into InferWidths
